[Spring 지하철 노선도 관리 - 3단계] 크로플(이동환) 미션 제출합니다. (#146)
* refactor() : Valid를 통한 예외 처리 및 패키지 조정

@NotBlank와 @Pattern을 이용해 생긴 예외를 handleValidationException을 통해 핸들링

* feat(Section) : Line insert 시 Section도 DB에 insert 하도록 구현 및 메서드 별 Dto 분리

* feat(LineInsertRequest) : Section 으로 변환하는 메서드 추가

* feat(LineService) : 노선 추가 시 Section 추가하고 해당 station들을 정렬해서 반환

* test() : 테스트에 필요한 변수들과 create 같은 메소드를 추출함

@DirtiesContext를 지우고 @Sql을 통해 테이블 초기화해서 테스트들의 독립성 보장

* feat(Section) : 구간 추가 기능 구현

* feat(Section) : 구간 삭제 기능 구현 및 구간 추가 시 거리 update 하도록 변경

* style() : code formatting

* refactor(LineService) : station과 section DAO 대신 service를 의존하도록 변경

* feat(StationService) : station 삭제할 때 section에 존재하고 있다면 예외 발생

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,9 +2,11 @@ package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.station.Station;
-import wooteco.subway.dto.response.StationResponse;
+import wooteco.subway.domain.Station;
+import wooteco.subway.dto.station.response.StationResponse;
+import wooteco.subway.exception.station.StationDeleteException;
 import wooteco.subway.exception.station.StationDuplicateException;
+import wooteco.subway.exception.station.StationNotExistException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -13,9 +15,11 @@ import java.util.stream.Collectors;
 public class StationService {
 
     private final StationDao stationDao;
+    private final SectionService sectionService;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionService sectionService) {
         this.stationDao = stationDao;
+        this.sectionService = sectionService;
     }
 
     public StationResponse create(String name) {
@@ -24,7 +28,6 @@ public class StationService {
                     throw new StationDuplicateException(name);
                 });
         return new StationResponse(stationDao.insert(name));
-
     }
 
     public List<StationResponse> showAll() {
@@ -34,7 +37,15 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
+    public Station showById(Long stationId) {
+        return stationDao.findById(stationId)
+                .orElseThrow(() -> new StationNotExistException(stationId));
+    }
+
     public void deleteById(Long id) {
+        if (sectionService.existStationByStationId(id)) {
+            throw new StationDeleteException(id);
+        }
         stationDao.deleteById(id);
     }
 }
